CP-54351 move cbtutil from /opt
